modified js file and created pages where needed
