+++ assets/js/components/sidebar/sidebar.html
@@ -20,15 +20,15 @@
 				
 			</a>
 		</li>
-		<li class="nav-item mt-2">
+		<!-- <li class="nav-item mt-2">
 			<a href="/pages/individual-contributions.html" class="nav-link rounded-pill">
-				<!--BUG: Cannot insert SVG here-->	
-				<img class = "nav-link-icon" src ="../../../../assets/images/icons/user-solid.svg" alt = "An icon representing a user"/><span>Individual Contributions</span>						
+				BUG: Cannot insert SVG here-->	
+				<!-- <img class = "nav-link-icon" src ="../../../../assets/images/icons/user-solid.svg" alt = "An icon representing a user"/><span>Individual Contributions</span>						 -->
 				
-			</a>
-		</li>
+			<!-- </a> -->
+		<!-- </li> --> 
 		<li class="nav-item mt-2">
-			<a href="" class="nav-link rounded-pill">
+			<a href="/pages/group_stats.html" class="nav-link rounded-pill">
 				<!--BUG: Cannot insert SVG here-->	
 				<img class = "nav-link-icon" src ="../../../../assets/images/icons/chart-simple-solid.svg" alt = "An icon representing a bar chart"/><span>Group Stats</span>
