+++ pages/activity-tracker.html
@@ -33,7 +33,7 @@
             crossorigin="anonymous"></script>
         <script src="https://cdn.jsdelivr.net/npm/chart.js@3.2.1/dist/chart.min.js"></script>
         <script src="../assets/js/components/component-map.js" type="module" defer></script>
-        <script src="../assets/js/activity-tracker.js" type="module"></script>
+        <script src="../assets/js/dashboard.js" type="module"></script>
     </body>
 
 </html>
